+++ Applications/Analyze/test/testAnalyzeTool.cpp
@@ -35,6 +35,7 @@ using namespace OpenSim;
 using namespace std;
 
 void testTutorialOne();
+void testActuationAnalysisWithDisabledForce();
 
 // Test different default activations are respected when activation
 // states are not provided.
@@ -68,6 +69,12 @@ int main()
     catch (const std::exception& e) {
         cout << e.what() << endl;
         failures.push_back("testTugOfWar with activation state provided");
+    }
+
+    try { testActuationAnalysisWithDisabledForce(); } 
+    catch (const std::exception& e) {
+        cout << e.what() << endl;
+        failures.push_back("testActuationAnalysisWithDisabledForce");
     }
 
     if (!failures.empty()) {
@@ -226,4 +233,25 @@ void testTugOfWar(const string& dataFileName, const double& defaultAct) {
         SimTK_ASSERT_ALWAYS(delta < maxDelta,
             "Force trajectory has unexplained discontinuity.");
     }
+}
+
+void testActuationAnalysisWithDisabledForce() {
+    AnalyzeTool analyze("PlotterTool.xml");
+    Model& model = analyze.getModel();
+    auto& muscle = model.updMuscles()[0];
+    muscle.set_appliesForce(false);
+
+    std::string resultsDir = "testPlotterToolWithDisabledForce";
+    analyze.setResultsDir(resultsDir);
+    analyze.run();
+
+    // Reading a file with mismatched nColumns header and actual number of
+    // data columns will throw.
+    TimeSeriesTable_<double> act_force_table(
+            resultsDir + "/BothLegs_Actuation_force.sto");
+    
+    // Let's also check that the number of columns is correct (i.e.,
+    // (number of muscles in the model) - 1).
+    ASSERT_EQUAL(model.getMuscles().getSize() - 1, 
+            (int)act_force_table.getNumColumns());
 }