+++ CHANGELOG.md
@@ -7,6 +7,7 @@ request related to the change, then we may provide the commit.
 This is not a comprehensive list of changes but rather a hand-curated collection of the more notable ones. For a comprehensive history, see the [OpenSim Core GitHub repo](https://github.com/opensim-org/opensim-core).
 v4.3
 ====
+- Fixed a bug with Actuation analysis that would lead to extra columns in the output when an actuator is disabled (Issue #2977).
 
 v4.2
 ====